Better orchestrate configuration steps
Three separate execution environments are needed.

The first one is from the Live ISO and installs the system

The second one is the chrooted environment that configures booting,
users, and network.  This is critical to ensure that the operations in
the next environment succeed.

The third one is the boot of the minimal system that allows systemd
operations to succeed, such as setting time and locale.

Update the installation process to work with NVMe drives.

More accurately set locale for all instances.

Ensure reflector is able to run and has all inputs.