Add configurations for LUKS-encrypted root and drop OPAL
Enabling TCG OPAL 2.0 on an NVMe SSD using sedutil and trying to boot
off of it with Linux is a mess.  More generally, using OPAL has always
been tricky and fraught with challenges and surprises.

In the past, SSD manufacturers have disclosed vulnerabilities and have
recommended using software-based encryption over their hardware-based
encryption.

Additionally, SSD and motherboard manufacturers need to coordinate to
get the BIOS/UEFI firmware to recognize all types of media for booting
where TCG OPAL 2.0 can be used as storage mediums evolve.  This is often
not done fast enough before products come to market.

Getting a working setup requires extensive research to determine which
motherboard and SSD combinations work.  This alone is a hassle and
becomes limiting when selecting desired motherboards and SSDs.

Then, once enabled, making any changes requires a careful unlocking
process.  This sometimes includes flashing a new BIOS/UEFI firmware, and
always includes anytime the OS needs to be reinstalled.

Security is worthwhile as long as it doesn't come at a huge
inconvenience cost as what's seen with OPAL.  Nowadays, CPUs have
dedicated hardware to accelerate encryption.  LUKS likely has this
behind it from the Linux kernel.  From a performance perspective, the
difference from a hardware setup is negligible.

This software setup with LUKS works with most hardware combinations and
is painless to set up.  This already makes it a superior alternative to
OPAL.

From a security perspective, the boot partition is still unencrypted.
This is an acceptable trade-off in favor of encrypting the bulk of the
space where sensitive data may be stored.  The idea here is to create a
deterrent to prevent access to the data in case of theft or loss and
does not have to be perfect.

While maintaining a separate home partition is nice, it's no longer
required.  Any time work is being done that has a chance at touching
home, data should be backed up anyway.  Having the option to quickly
wipe and restore everything to apply new configurations (e.g. new file
system) without needing to fiddle much with the scripts is the important
part of what this enables.

This also requires using a swap file rather than a dedicated swap
partition.  This is more flexible.  Support for this will be added back.

Need to disable hibernate support for now until can figure out how to
get this working with encryption.