Better set PS1
PS1 is ultimately a user decision, but want to provide a better default.
Do this by modifying the bashrc skeleton so that it gets created for
each user on first login.

Add the installation of gnome-terminal to this script because it makes
sense with the extra vte customization, which is only needed for
GTK-based terminal emulators.