Switch entirely to open source fonts
Use Noto Sans for sans-serif characters because it is a humanist font
and given its wide Unicode coverage, it eliminates tofu square boxes due
to missing glyphs on the web.

Use Cascadia Code PL for monospace characters because the typeface is
enjoyable to use when programming and contains Powerline glyphs.

Use FreeSerif for serif characters because it is very close to Times New
Roman.

Only intall ttf-ms-fonts for broad web rendering compatibility for sites
that expect these fonts.