Switch back to X11 as the default display server protocol
Things continue to just work better on X11.  The user interface is
scaled correctly on the login screen when using a HiDPI monitor.  The
login screen correctly appears on the designated primary monitor.  All
windows stay associated with their icon on the taskbar.  Tresorit is
working again to back up files securely.  Visual Studio Code has window
shadows again.  And so far with X11, the HDMI connections to monitors
appear to be stable without periodic loss of signal.