Define XDG user directory variables
This should not be needed for programs that adhere to the XDG spec,
since these variables are set to the default fallback values. For
programs that do not adhere to the XDG spec, they may benefit from
redundant definition of these variables.

This won't get everything, but should help to organize files and folders
that appear in the user's home directory better. It should also make
following online documentation easier when values for these variables
are actually set.