Add Unreal Engine 5 setup script
It's a pipe dream to do work with this.  Needing to prioritize a clean
home office setup that can seamlessly switch KVM + other devices between
home and work computers to minimize crazy wire hell is more important at
the moment.

Though the monitors no longer have VRR and are now susceptible to
tearing again, development is still possible and it might be picked up
one day.