No longer modify vm.swappiness
Believe that default values are set for a reason

Also, decreasing swappiness prefers to keep programs in memory, which
when full, makes new programs take a performance hit when unused
programs should be swapped out of RAM and to the disk to make room in
RAM for the new and active programs.

This also doesn't really matter anymore since there is no swap.  Setting
it up is complicated on a LUKS-encrypted partition with a swap file, and
having a dedicated and encrypted partition for swap is even more complex
to set up.  Also, swap is just not needed in this case due to the large
amount of RAM in the system.