Switch to KDE
Prefer to use a desktop that is all about the freedom of choice and
customization.

The desktop is the way I want it out of the box because it provides the
traditional desktop metaphor.  It doesn't need the help of extensions
just to become usable.

I like the community surrounding KDE more than that of GNOME, which
sometimes thinks they know best for the user and has, in a number of
instances, been hand-waving and dismissive.  It's still like this, which
is not inviting.

As a part of this switch, begin using more KDE-centric or desktop
environment-independent applications, such as virt-manager.

SDDM is slated to become an official part of KDE.  Use this over GDM
from GNOME.

Remove GNOME-specific settings, make generic for KDE to get GTK apps to
integrate well into the default theme