Revise primary package list
Use Thunderbird as a mail client.  This is sorely needed since mail,
calendar, and contacts have gotten out of control :)  The refreshed UI
in Thunderbird is quite nice.

Drop C++ development tools, since these are not the primary focus of
recreational activities at the moment.

Use Haruna as a MPV player skin, since it is Qt-based and integrates
better with KDE than the GTK-based equivalent of Celluoid.

Drop unneded utilities now that KDE provides a good set of functionality
out of the box.