Expand and reorganize user package installation
Generally break down into three sections:

* Primary user applications
* AUR user applications
* Base desktop applications

Then create subcategories based on those that exist in the
kde-applications-meta package.

Operationally, this makes it easy to track where applications come from.  It
puts more of a focus on what the user initially cares about.  It shows exactly
what is coming from the community at a glance.  And it shows what supporting
auxiliary but important programs that are expected to come with the default
environment.

Switch to Krita over Pinta to avoid having pacman manage a .NET Core
installation.  This is now done by the user.

Add packages that came in for free with the GNOME meta package, but now using
the minimal plasma meta package.  List exactly what is needed to avoid bringing
in the world of KDE, which is quite large.

Isolate AUR commands to only the packages that need them.

Minimize the packages that get installed during the chroot step, since we want
to get going to a nearly installed terminal ASAP.