Work better with default Phonon backend and Wayland for multimedia
The plasma-meta package chooses the phonon-qt5-gstreamer backend by
default in an unattended install.  This doesn't install an extra video
player, which is nice.  Work with this default and fully install all
parts of the gstreamer infrastructure and ensure hardware video
acceleration is enabled.

Switch to haruna for the media player since it works very well with
Wayland.