Switch back to NVIDIA graphics
Dual booting with Windows and Linux on different physical drives.  Using
Windows to complement gaming on the PS5 to get access to exclusive Xbox
titles that are also released on the PC.  Need the absolute best in
performance here because will be playing the Ori games at their highest
quality settings on a 4K television at 120Hz refresh rate over HDMI 2.1.
Want to use a graphics card with native HDMI support to avoid any
compatibility issues with a television that is not certified as G-SYNC
compatible.  Intel is using a protocol conversion chip in their Arc
Alchemist lineup:

https://game.intel.com/story/intel-arc-graphics-vrr-hdr-hdmi

This is more tolerable now on the Linux side since NVIDIA recently
(finally) implemented support for GAMMA_LUT, to enable night light in
GNOME/KDE.