Enable mode to preserve home
Preserve /home by default, since we have settled on a partitioning and
encryption scheme.  It's also generally good practice to make
destructive operations explicit.  Warn and give the user 5 seconds to
abort the destructive operation to make sure they really want this.

Want to be able to reinstall cleanly (to take care of stateful
directories like /etc and /var), but not need to backup data locally,
which is an annoying operation.  Data is already backed up remotely ;)

This can be done by taking advantage of the separate paritions and
logical volume groups that exist.  Everything except for /home can be
overwritten.

Move to getopts since this is a little more concise when parsing
arguments.