Allow crypt device time to show up
Preserving the /home partition didn't work quite as expected initially.  It
complained that the volume group didn't exist.  When re-running the script, it
then complained that the crypt device was already open.  After that was fixed,
everything just magically worked without any other changes.  When not creating
volumes explicitly, there must be a small delay in the registration of the
devices that needs to be accounted for.