Default package changes
Firefox now comes with Wayland enabled by default, so no need to handle this
manually anymore

Prefer to integrate best with GNOME by using Contacts, Calendar, and Geary for
PIM instead of using Thunderbird

Switch from pinta to gimp-git to have dotnet be entirely user-managed and avoid
adding a dependency on GTK2

Drop LibreOffice in favor of editing documents on Windows with Office.  This
has always been a pain point of using Linux, and given the current dual-boot
setup with Windows for gaming, it makes sense to also use it for Office-related
work.