Rework installed applications
Keep the AUR section strictly for installing packages sourced from the AUR

No longer install Arch Linux artwork.  It was only used for the application
launcher as a part of the Dash to Panel GNOME extension, but this has since
been replaced with the ArcMenu which comes with its own icons that are colored
better.

Remove Hugo.  Not interested in maintaining a blog.  Would rather let projects
on GitHub speak for themselves.

Remove QCad.  Not messing around with the idea of replacing counter tops
anymore.

Replace Celluoid and the FFmpeg thumbnailer with GNOME Videos, which comes with
its own thumbnailer, for video playback.  Use GNOME Music for audio playback.
Now that we have video encoding and decoding working well on modern Intel
hardware, can use applications that are more integrated into the GNOME desktop
environment/software stack.

Add HandBrake CLI to aid in more easily automating video conversion