Revert "Switch back to NVIDIA graphics"
This reverts commit e2c4a71711ec44f30f4808b8bbcb34adbe1331a7.

Video encoding with the NVIDIA RTX 3070 is 5.79% faster than the Intel ARC
A770.  The 3070 also performs better with rendering in Unreal Engine 5.

However, these are not primary use cases.  The A770 integrates better with the
Linux stack, has better day-to-day rendering still, though the gap is closing
now that explicit sync is working.

For now, stick with Intel :)