Revert "Reapply "Switch back to NVIDIA graphics""
This reverts commit f65dad3debfa069f08b97987009751b216050f08.

Using NVIDIA graphics is the right move for having a gaming PC.  However, I
expect gaming to be done in the living room, and carting a PC up and down the
stairs every time I want to play a game isn't worth it.  I'd need to buy a
laptop as my daily driver for this to work, and that's not happening right now.
I'll keep using the PS5 for the time being, and one day transition onto PC
gaming.

For now, switch back to Intel graphics for the best possible Linux experience.