Switch back to NVIDIA graphics
The latest Intel driver cannot be used to launch Unreal Engine 5.5, which has
high requirements around Vulkan support and no longer supports OpenGL.  This
issue is not present in the Nvidia driver.

Additionally, video encoding times and file sizes are slightly better with
Nvidia.  And, there is a proper HDMI port on the card, which makes for
fantastic HDR and automatic ALLM engagement on modern TVs.

The only place where Intel wins is the VRAM and Wayland support.  Both are
important, but are not what ultimately matter at the end of the day when you
get a GPU to DO stuff.