Streamline system utilities
GParted is a superior GUI for partitioning and formatting disks; replaces GNOME
Disk Utility.  Mission Center is a superior GUI for monitoring performance and
applications; replaces GNOME System Monitor.

Add CLI disk utilities for SMART monitoring that are similar to hdparm and
nvme-cli.

Remove leftover Hardinfo2 which is largely not used.